test_requirements.txt
django>=1.11
djangorestframework>=3.9
git+https://github.com/AlbinLindskog/django_tokens
